Define cli helpers in ckan.cli.cli
