+++ ckan/cli/cli.py
@@ -3,13 +3,18 @@
 import logging
 from collections import defaultdict
 
+import six
 import click
-from ckan.cli import config_tool
+
+import ckan.plugins as p
+from ckan.config.middleware import make_app
 from ckan.cli import (
+    load_config,
+    config_tool,
     jobs,
     datapusher,
     front_end_build,
-    click_config_option, db, search_index, server,
+    db, search_index, server,
     profile,
     asset,
     datastore,
@@ -29,6 +34,37 @@ from ckan.cli import (
 from ckan.cli import seed
 
 log = logging.getLogger(__name__)
+
+
+class CkanCommand(object):
+
+    def __init__(self, conf=None):
+        self.config = load_config(conf)
+        self.app = make_app(self.config.global_conf, **self.config.local_conf)
+
+
+def _init_ckan_config(ctx, param, value):
+    ctx.obj = CkanCommand(value)
+    if six.PY2:
+        ctx.meta["flask_app"] = ctx.obj.app.apps["flask_app"]._wsgi_app
+    else:
+        ctx.meta["flask_app"] = ctx.obj.app._wsgi_app
+
+    for plugin in p.PluginImplementations(p.IClick):
+        for cmd in plugin.get_commands():
+            cmd._ckanext = plugin.name
+            ctx.command.add_command(cmd)
+
+
+click_config_option = click.option(
+    u'-c',
+    u'--config',
+    default=None,
+    metavar=u'CONFIG',
+    help=u'Config file to use (default: development.ini)',
+    is_eager=True,
+    callback=_init_ckan_config
+)
 
 
 class CustomGroup(click.Group):
